# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GIQ HR Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job postings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, open, closed, archived]
        - name: department
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: employmentType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: public
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /analytics/activity:
    get:
      operationId: getActivityTrends
      tags: [analytics]
      summary: Get hiring activity trends
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Activity trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityTrendsResponse"

  # Notifications
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, recruiter]
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
        - updatedAt
...
    Job:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contract, internship]
        seniority:
          type: string
          enum: [entry, mid, senior, lead, executive]
        requiredSkills:
          type: array
          items:
            type: string
        preferredSkills:
          type: array
          items:
            type: string
        minExperience:
          type: integer
          nullable: true
        maxExperience:
          type: integer
          nullable: true
        minSalary:
          type: integer
          nullable: true
        maxSalary:
          type: integer
          nullable: true
        salaryCurrency:
          type: string
          nullable: true
        description:
          type: string
        responsibilities:
          type: string
        qualifications:
          type: string
        status:
          type: string
          enum: [draft, open, closed, archived]
        createdById:
          type: string
          nullable: true
        createdBy:
          $ref: "#/components/schemas/User"
          nullable: true
        candidateCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - department
        - location
        - employmentType
        - seniority
        - requiredSkills
        - preferredSkills
        - description
        - responsibilities
        - qualifications
        - status
        - candidateCount
        - createdAt
        - updatedAt
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - jobs
        - meta
    CreateJobBody:
      type: object
      required: [title, department, location, employmentType, seniority, description, responsibilities, qualifications]
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contract, internship]
        seniority:
          type: string
          enum: [entry, mid, senior, lead, executive]
        requiredSkills:
          type: array
          items:
            type: string
          default: []
        preferredSkills:
          type: array
          items:
            type: string
          default: []
        minExperience:
          type: integer
          nullable: true
        maxExperience:
          type: integer
          nullable: true
        minSalary:
          type: integer
          nullable: true
        maxSalary:
          type: integer
          nullable: true
        salaryCurrency:
          type: string
          nullable: true
        description:
          type: string
        responsibilities:
          type: string
        qualifications:
          type: string
        status:
          type: string
          enum: [draft, open, closed, archived]
          default: draft
...
    ActivityDataPoint:
      type: object
      properties:
        date:
          type: string
        candidatesAdded:
          type: integer
        interviewsScheduled:
          type: integer
        screeningsCompleted:
          type: integer
      required:
        - date
        - candidatesAdded
        - interviewsScheduled
        - screeningsCompleted
    ActivityTrendsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActivityDataPoint"
      required:
        - data
...